third_party/blink/web_tests/external/wpt/mathml/presentation-markup/tables/table-default-styles-001-expected.txt

This is a testharness.js-based test.
PASS Default math-style value on mtable
FAIL Default display value on mtable assert_equals: expected "inline-table" but got "table"
PASS Default display value on mtr
PASS Default display value on mtd
FAIL Default text-align value on mtd assert_equals: expected "center" but got "-webkit-center"
PASS Default padding-left value on mtd
PASS Default padding-right value on mtd
PASS Default padding-top value on mtd
PASS Default padding-bottom value on mtd
PASS Bounding box is the same as mrow + mtd default padding
Harness: the test ran to completion.
